AI_Text/requirements.txt
fastapi==0.103.0
pydantic==1.10.9
openai==0.28.0
python-dotenv==1.0.0
uvicorn==0.31.0
